3916 - Cycle clone: metadata (refactor)
