+++ src/server/controller/assessment/cloneCycle/types.ts
@@ -1,0 +1,7 @@
+import { Assessment, Cycle } from 'meta/assessment'
+
+export type CloneProps = {
+  assessment: Assessment
+  cycleSource: Cycle
+  cycleTarget: Cycle
+}
